Move FlagEnum into kernel_api.
