+++ numba_dpex/kernel_api/flag_enum.py
@@ -2,7 +2,8 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-"""Provides a FlagEnum class to help distinguish IntEnum types that numba-dpex
+"""
+Provides a FlagEnum class to help distinguish IntEnum types that numba_dpex
 intends to use as Integer literal types inside the compiler type inferring
 infrastructure.
 """
@@ -10,13 +11,13 @@ from enum import IntEnum
 
 
 class FlagEnum(IntEnum):
-    """Helper class to distinguish IntEnum types that numba-dpex should consider
+    """Helper class to distinguish IntEnum types that numba_dpex should consider
     as Numba Literal types.
     """
 
     @classmethod
     def basetype(cls) -> int:
-        """Returns an dummy int object that helps numba-dpex infer the type of
+        """Returns an dummy int object that helps numba_dpex infer the type of
         an instance of a FlagEnum class.
 
         Returns:
