+++ numba_dpex/tests/experimental/codegen/test_intenum_literal_codegen.py
@@ -9,7 +9,7 @@ from numba.core import types
 
 import numba_dpex.experimental as exp_dpex
 from numba_dpex import DpctlSyclQueue, DpnpNdArray, int64
-from numba_dpex.experimental.flag_enum import FlagEnum
+from numba_dpex.kernel_api.flag_enum import FlagEnum
 
 
 def test_compilation_as_literal_constant():
